Fix issue with extracting files

+++ core/src/main/java/com/ryderbelserion/fusion/core/api/utils/FileUtils.java
@@ -114,7 +114,9 @@ public class FileUtils {
                         Files.createDirectories(parent);
                     }
 
-                    Files.copy(stream, target);
+                    if (!Files.exists(target)) {
+                        Files.copy(stream, target);
+                    }
                 }
             }
         } catch (final URISyntaxException | IOException exception) {
